feat: Auth module documentation
