+++ api.planx.uk/docs/parameters.yaml
@@ -1,0 +1,25 @@
+openapi: 3.1.0
+info:
+  title: Plan✕ API
+  version: 0.1.0
+paths: {}
+components:
+  parameters:
+    sessionId:
+      in: path
+      name: sessionId
+      type: string
+      required: true
+      description: Session ID
+    localAuthority:
+      in: path
+      name: localAuthority
+      type: string
+      required: true
+      description: "Name of the Local Authority, usually the same as Planx `team`"
+    hasuraAuth:
+      name: authorization
+      in: header
+      descriptioN: An authorisation header provided by Hasura
+      required: true
+      type: string
