+++ api.planx.uk/server.ts
@@ -12,8 +12,6 @@ import { Server } from "http";
 import passport from "passport";
 import helmet from "helmet";
 import multer from "multer";
-import swaggerJSDoc from "swagger-jsdoc";
-import swaggerUi from "swagger-ui-express";
 
 import { ServerError } from "./errors";
 import { locationSearch } from "./gis/index";
@@ -79,84 +77,13 @@ import { createPaymentSendEvents } from "./inviteToPay/createPaymentSendEvents";
 import { getSessionSummary } from "./admin/session/summary";
 import { googleStrategy } from "./modules/auth/strategy/google";
 import authRoutes from "./modules/auth/routes";
+import { useSwaggerDocs } from "./docs";
 
 const router = express.Router();
 
 const app = express();
 
-// Swagger documentation config
-const options = {
-  failOnErrors: true,
-  definition: {
-    openapi: "3.1.0",
-    info: {
-      title: "Plan✕ API",
-      version: "0.1.0",
-      description:
-        "Plan✕ is a platform for creating and publishing digital planning services. Our REST API is built on Express and documented with Swagger.",
-      license: {
-        name: "MPL-2.0",
-        url: "https://github.com/theopensystemslab/planx-new/blob/main/LICENSE.md",
-      },
-    },
-    schemes: ["http", "https"],
-    servers: [{ url: process.env.API_URL_EXT }],
-    components: {
-      securitySchemes: {
-        bearerAuth: {
-          type: "http",
-          scheme: "bearer",
-          bearerFormat: "JWT",
-        },
-      },
-      parameters: {
-        sessionId: {
-          in: "path",
-          name: "sessionId",
-          type: "string",
-          required: true,
-          description: "Session ID",
-        },
-        localAuthority: {
-          in: "path",
-          name: "localAuthority",
-          type: "string",
-          required: true,
-          description:
-            "Name of the Local Authority, usually the same as Planx `team`",
-        },
-        hasuraAuth: {
-          name: "authorization",
-          in: "header",
-          descriptioN: "An authorisation header provided by Hasura",
-          required: true,
-          type: "string",
-        },
-      },
-      schemas: {
-        SessionPayload: {
-          type: "object",
-          properties: {
-            payload: {
-              type: "object",
-              properties: {
-                sessionId: {
-                  type: "string",
-                },
-              },
-            },
-          },
-          example: {
-            payload: { sessionId: "123" },
-          },
-        },
-      },
-    },
-  },
-  apis: ["./**/*.ts", "./**/*.js"],
-};
-const swaggerSpec = swaggerJSDoc(options);
-app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerSpec));
+useSwaggerDocs(app);
 
 app.set("trust proxy", 1);
